Some time between when I committed last and when I checked out today, something was checked in that broke my compile.  Minimap.o was no longer in makefile.linux, yet some of its functions were needed.  Also, someone defined a non-static, non-const global in minimap.h (don't do that!  ;)).
Fixed, at least for me.

+++ eye_candy/eye_candy.cpp
@@ -696,7 +696,7 @@ coord_t Particle::flare() const
   assert(flare_frequency);
   if (flare_max == 1.0)
     return 1.0;
-  const int offset = long(&alpha);	//Unique to the particle.
+  const short offset = (short)long(&alpha);	//Unique to the particle.
   math_cache.powf_0_1_rough_close(fabs(sin((pos.x + pos.y + pos.z + offset) / flare_frequency)), flare_exp);
   const coord_t flare_val = 1.0 / (math_cache.powf_0_1_rough_close(fabs(sin((pos.x + pos.y + pos.z) / flare_frequency + offset)), flare_exp));
   if (flare_val > flare_max)
